Перенес функцию parseFormData из form-edit в controllers/task.js и откоректировал работу с этой функцией
